minor positioning adjustments

+++ src/xeed/library/common/Utils.java
@@ -5,8 +5,8 @@ import android.util.TypedValue;
 
 public final class Utils
 {
-    public static final int getPx(final Context c, final float dp)
+    public static final float getPx(final Context c, final float dp)
     {
-        return (int)TypedValue.applyDimension(TypedValue.COMPLEX_UNIT_DIP, dp, c.getResources().getDisplayMetrics());
+        return TypedValue.applyDimension(TypedValue.COMPLEX_UNIT_DIP, dp, c.getResources().getDisplayMetrics());
     }
 }
